Добавлена возможность использовать внешнюю файловую систему

+++ internal/fileStorage/filesystem/export.go
@@ -11,6 +11,10 @@ import (
 func (s *Storage) CreateExportFile(ctx context.Context, name string) (io.WriteCloser, error) {
 	defer system.Stopwatch(ctx, "CreateExportFile")()
 
+	if s.readOnly {
+		return nil, readOnlyModeError
+	}
+
 	f, err := os.Create(path.Join(s.exportPath, name))
 	if err != nil {
 		system.Error(ctx, err)
